Begin making the code generic for any kind of stream.
There is still more that can be done, but my knowledge of Modbus isn't extensive.
For instance, MODBUS_PROTOCOL_TCP is hardcoded, and I assume that a complete
generic implementation must provide a way to abstract that, too.

I've also made the API object-oriented by turning functions like `read_coils` into methods
on `Context`.
